make keyboard disappear on touch outside of keyboard
